ipc_send_event(), ipc_send_client_message(): take json_object
Let these functions take a json_object instead of a sequence of bytes.
Keep old versions of these functions but append `_raw` to their names.
This change prepares for migrating from `yajl` to `json-c`.

In `load_layout.c`, rename `json_bool` to `json_boolean` to avoid name
conflict with a typedef in `json-c`.